moved code to operators
